% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ma.R
\name{calc_ma}
\alias{calc_ma}
\title{Calculate mass accuracy}
\usage{
calc_ma(m, m_cal, ...)
}
\arguments{
\item{m}{Measured mass}

\item{m_cal}{Calculated (theoretical) mass.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A numeric vector of mass accuracy (rounded to 4 decimals).
}
\description{
Calculates relative mass accuracy (ma, in parts per million) as:

\eqn{(m_{meas} - m_{calc}) / m_{calc} \times 10^6}
where:
\itemize{
\item \eqn{m_{meas}} = measured mass
\item \eqn{m_{calc}} = calculated / theoretical (exact) mass
}
Returned value is rounded to 4 digits.
In this context the theoretical mass is represented by the mass
of the assigned molecular formula.
A small absolute \strong{ppm value} indicates a very precise measurement and increases
confidence in correct molecular formula assignment.
}
\examples{
# Use of single values
calc_ma(m = 264.08641, m_cal = 264.08653)
# Use in a molecular formula table
calc_ma(m = mf_data_demo$m, m_cal = mf_data_demo$m_cal)
mf_data_demo[, .(m, m_cal, accuracy_in_ppm = calc_ma(m, m_cal))]
}
\seealso{
Other calculations: 
\code{\link{calc_data_summary}()},
\code{\link{calc_dbe}()},
\code{\link{calc_eval_params}()},
\code{\link{calc_exact_mass}()},
\code{\link{calc_ideg}()},
\code{\link{calc_neutral_mass}()},
\code{\link{calc_nm}()},
\code{\link{calc_norm_int}()},
\code{\link{calc_number_assignment}()},
\code{\link{calc_number_occurrence}()},
\code{\link{calc_recalibrate_ms}()}
}
\concept{calculations}
\keyword{misc}
