% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ma.R
\name{calc_ma_abs}
\alias{calc_ma_abs}
\title{Calculate absolute mass accuracy range (ma)}
\usage{
calc_ma_abs(m, ma_dev, ...)
}
\arguments{
\item{m}{Measured mass}

\item{ma_dev}{Mass accuracy in +/- parts per million (ppm)}

\item{...}{Additional arguments passed to methods.}
}
\value{
Returns a list with two values: m_min, m_max
}
\description{
This function calculates the absolute mass accuracy range for a neutral mass (m) at a given a mass accuracy (ma_dev).
}
\examples{
calc_ma_abs(m = 327.0134, ma_dev = 0.5)
}
\keyword{misc}
