% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_objects.R
\name{check_mfd}
\alias{check_mfd}
\title{Check format of molecular formula data}
\usage{
check_mfd(mfd, ...)
}
\value{
A \code{data.table} containing the validated and standardized molecular formula
data. The function checks column names, ensures the presence of essential
variables (\code{file_id}, \code{mz}, \code{m}, \code{ppm}), renames isotope columns when needed,
and adds missing columns if necessary. The returned \code{data.table} is the input
object \code{mfd}, potentially modified in place.
}
\description{
Verify the correct usage of UME column names, existence of a unique peak identifier (peak_id),
and a unique file/analysis name (file_id).
Remove rows having missing values for either m/z (mz) or peak magnitude (i_magnitude).
}
\seealso{
Other check ume objects: 
\code{\link{as_peaklist}()},
\code{\link{check_formula_library}()}
}
\concept{check ume objects}
\keyword{internal}
