% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_objects.R
\name{check_table_schema}
\alias{check_table_schema}
\title{Check data.table structure}
\usage{
check_table_schema(dt, schema, name = "table")
}
\arguments{
\item{dt}{A data.table to check.}

\item{schema}{A schema list object as defined in \verb{.ume_schema_*}.}

\item{name}{Optional: name of the table (for clearer error messages)}
}
\value{
Logical TRUE/FALSE invisibly.
}
\description{
Internal helper to verify if a table matches a defined ume schema.
}
\keyword{internal}
