% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.extract_library_version}
\alias{.extract_library_version}
\title{Extract UME library version from formula library object}
\usage{
.extract_library_version(lib)
}
\arguments{
\item{lib}{A formula library data.table or list.}
}
\value{
Numeric library version.
}
\description{
Extract UME library version from formula library object
}
\keyword{internal}
