% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_schemas.R
\docType{data}
\name{.palette_builders}
\alias{.palette_builders}
\title{CENTRAL PALETTE REGISTRY}
\format{
An object of class \code{list} of length 9.
}
\usage{
.palette_builders
}
\description{
defines all palettes in ONE place.
}
\keyword{internal}
