% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_schemas.R
\docType{data}
\name{.ume_schema_peaklist}
\alias{.ume_schema_peaklist}
\title{Data table schemas used in ume}
\format{
An object of class \code{list} of length 2.
}
\usage{
.ume_schema_peaklist
}
\description{
Internal definitions of expected column structures for key ume table types.
}
\keyword{internal}
