% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_ume_peaklist}
\alias{is_ume_peaklist}
\title{Check whether an object is a UME peaklist}
\usage{
is_ume_peaklist(x)
}
\arguments{
\item{x}{Any object}
}
\value{
TRUE/FALSE
}
\description{
Check whether an object is a UME peaklist
}
