% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lib_demo}
\alias{lib_demo}
\title{Demo formula library (200 - 300 Da, neutral mass)}
\format{
A data.table having ~115,111 rows and 12 variables:
\describe{
\item{vkey}{First two digits represent the formula library version;
last digits are unique identifiers for each formula}
\item{mf}{Neutral molecular formula (no differentiation of isotopes)}
\item{mass}{Calculated exact neutral mass of a formula (based on ume::masses)}
}
}
\usage{
lib_demo
}
\description{
Contains a small molecular formula library for demonstration and
validation purposes. Complete formula libraries are available
in the 'ume.formulas' data package.
}
\examples{
data(peaklist_demo)
}
\seealso{
Other ume data: 
\code{\link{known_mf}},
\code{\link{masses}},
\code{\link{mf_data_demo}},
\code{\link{nice_labels_dt}},
\code{\link{peaklist_demo}},
\code{\link{tab_ume_labels}}
}
\concept{ume data}
\keyword{datasets}
