% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nice_labels_dt}
\alias{nice_labels_dt}
\title{nice_labels_dt}
\format{
A data.table with labels that can be used for plots
\describe{
\item{name_substitute}{Name that will be displayed instead of the standard column name}
\item{name_pattern}{Name of the standard column in ume tables}
}
}
\source{
{taken from www.awi.de}
}
\usage{
nice_labels_dt
}
\description{
nice_labels_dt
}
\examples{
data(nice_labels_dt)
}
\seealso{
Other ume data: 
\code{\link{known_mf}},
\code{\link{lib_demo}},
\code{\link{masses}},
\code{\link{mf_data_demo}},
\code{\link{peaklist_demo}},
\code{\link{tab_ume_labels}}
}
\concept{ume data}
\keyword{datasets}
