% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaklist_demo}
\alias{peaklist_demo}
\title{Demo peak list}
\format{
A data.table having 31,091 rows and 7 variables:
\describe{
\item{file_id}{A unique identifier for a mass spectrum (integer)}
\item{file}{A unique label for a mass spectrum or sample (character)}
\item{peak_id}{A unique identifier for a peak in the entire peak list (integer)}
\item{mz}{Mass to charge ratio of the singly charged molecular ion (numeric)}
\item{i_magnitude}{Peak magnitude of the molecular ion (numeric)}
\item{s_n}{Signal to noise ratio of the molecular ion (numeric)}
\item{res}{Mass resolution of the peak / ion (numeric)}
}
}
\source{
{taken from www.awi.de}
}
\usage{
peaklist_demo
}
\description{
Contains parts of the peaklist (200 - 300 m/z) from mass spectra
to use as demonstration and validation dataset.
The sample mass spectra contain one blank, three replicates
of North Sea water, and three Arctic fjord samples as triplicates.
}
\examples{
data(peaklist_demo)
}
\seealso{
Other ume data: 
\code{\link{known_mf}},
\code{\link{lib_demo}},
\code{\link{masses}},
\code{\link{mf_data_demo}},
\code{\link{nice_labels_dt}},
\code{\link{tab_ume_labels}}
}
\concept{ume data}
\keyword{datasets}
