% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{remove_id_columns}
\alias{remove_id_columns}
\title{Remove columns that contain ID's}
\usage{
remove_id_columns(df, ...)
}
\arguments{
\item{df}{data.table that contains ID columns}

\item{...}{Additional arguments passed to methods.}
}
\description{
This functions removes columns ID columns ('_id') and hierarchical search columns ('_lft', '_rgt') from a table.
Only exceptions are "sample_id" and "bottle_id that are always kept in the output table.
}
\seealso{
Other Clean data output: 
\code{\link{remove_unknown_columns}()}
}
\concept{Clean data output}
\keyword{ID}
\keyword{Remove}
\keyword{columns.}
