% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{remove_unknown_columns}
\alias{remove_unknown_columns}
\title{Remove columns that only have one specific value}
\usage{
remove_unknown_columns(df, excl_cols = NULL, search_term = " unknown", ...)
}
\arguments{
\item{df}{data.table that contains empty columns}

\item{excl_cols}{List of column names that should not be removed, even if all values contain search_term}

\item{search_term}{String that uniquely occurs in one column}

\item{...}{Additional arguments passed to methods.}
}
\description{
This function removes columns that exclusively contain the value defined in 'search_term' (such as " unknown" (default)).
}
\seealso{
Other Clean data output: 
\code{\link{remove_id_columns}()}
}
\concept{Clean data output}
\keyword{Remove}
\keyword{columns}
\keyword{contain}
\keyword{exclusively}
\keyword{missing}
\keyword{that}
\keyword{values.}
