% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_objects.R
\name{revert_column_names}
\alias{revert_column_names}
\title{Revert data.table column names}
\usage{
revert_column_names(dt)
}
\arguments{
\item{dt}{data.table previously normalized with \code{normalize_columns}}
}
\value{
data.table with original column names restored
}
\description{
Restore the original column names recorded in \code{col_history}.
}
\keyword{internal}
