% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab_ume_labels}
\alias{tab_ume_labels}
\title{Labels of UME columns.}
\format{
A data.table that is derived from the MarChem database:
\describe{
\item{label}{Identifier for each label}
\item{nice_label}{Label that can be used e.g. in figures}
\item{use_in_ume}{Shows if label is used in the UME shiny app}
}
}
\source{
{taken from www.awi.de}
}
\usage{
tab_ume_labels
}
\description{
Labels of UME columns.
}
\examples{
data(tab_ume_labels)
}
\seealso{
Other ume data: 
\code{\link{known_mf}},
\code{\link{lib_demo}},
\code{\link{masses}},
\code{\link{mf_data_demo}},
\code{\link{nice_labels_dt}},
\code{\link{peaklist_demo}}
}
\concept{ume data}
\keyword{datasets}
