% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_uplots.R
\name{theme_uplots}
\alias{theme_uplots}
\title{theme_uplots}
\usage{
theme_uplots(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Numeric base font size.}

\item{base_family}{Font family.}
}
\value{
A ggplot2 theme object.
}
\description{
Applies a clean UME-style theme used across all uplot_* visualisations.
}
\details{
Unified UME Theme for All uplot_* Functions
}
