% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ume_logo_raster}
\alias{ume_logo_raster}
\title{Internal raster of the UME logo}
\format{
A 3D numeric array representing an RGBA raster image.
}
\description{
This object contains the preloaded raster version of the UME logo,
generated once during package development from
\code{inst/figures/ume_package_icon.png}.

It is used internally by \code{uplots_add_ume_logo()} to avoid runtime
dependencies on the \strong{png} package and to ensure the logo can be added
without file I/O.
}
\keyword{internal}
