% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplot_ms.R
\name{uplot_ms}
\alias{uplot_ms}
\title{Plot Mass Spectrum}
\usage{
uplot_ms(
  pl,
  mass = "mz",
  peak_magnitude = "i_magnitude",
  label = "file_id",
  logo = FALSE,
  plotly = TRUE,
  data_reduction = 1,
  ...
)
}
\arguments{
\item{pl}{A \code{data.table} containing at least columns for mass-to-charge ratio
and peak magnitude (e.g. a peak list or molecular formula data).}

\item{mass}{Character. Name of the column containing mass-to-charge or mass
information (default = \code{"mz"}).}

\item{peak_magnitude}{Character. Name of the column containing peak magnitude
(default = \code{"i_magnitude"}).}

\item{label}{Character. Name of the column identifying individual spectra
(default = \code{"file_id"}).}

\item{logo}{Logical. If TRUE, adds a UME caption.}

\item{plotly}{Logical. If TRUE, return interactive plotly object.}

\item{data_reduction}{Numeric between 0 and 1. Fraction of the most abundant
peaks to retain per spectrum. Default = 1 (no reduction).
If set to 0, a minimum of 0.01 is used to ensure some data is displayed.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{ggplot} object or a \code{plotly} object if \code{plotly = TRUE}.
}
\description{
Plots a mass spectrum, showing peak magnitude versus mass-to-charge ratio (m/z).

Optionally reduces the dataset by selecting the most abundant peaks per spectrum.
}
\examples{
uplot_ms(pl = peaklist_demo, data_reduction = 0.1, plotly = TRUE)
uplot_ms(pl = peaklist_demo, data_reduction = 1, plotly = FALSE)

}
\seealso{
Other uplots: 
\code{\link{uplot_cluster}()},
\code{\link{uplot_cvm}()},
\code{\link{uplot_dbe_minus_o_freq}()},
\code{\link{uplot_dbe_vs_c}()},
\code{\link{uplot_dbe_vs_ma}()},
\code{\link{uplot_dbe_vs_o}()},
\code{\link{uplot_freq_ma}()},
\code{\link{uplot_freq_vs_ppm}()},
\code{\link{uplot_hc_vs_m}()},
\code{\link{uplot_heteroatoms}()},
\code{\link{uplot_isotope_precision}()},
\code{\link{uplot_kmd}()},
\code{\link{uplot_lcms}()},
\code{\link{uplot_ma_vs_mz}()},
\code{\link{uplot_n_mf_per_sample}()},
\code{\link{uplot_pca}()},
\code{\link{uplot_ratios}()},
\code{\link{uplot_reproducibility}()},
\code{\link{uplot_ri_vs_sample}()},
\code{\link{uplot_vk}()}
}
\concept{uplots}
