% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplot_vk.R
\name{uplot_vk}
\alias{uplot_vk}
\title{uplot_vk}
\usage{
uplot_vk(
  mfd,
  z_var = "norm_int",
  projection = TRUE,
  palname = "viridis",
  median_vK = TRUE,
  col_median = "white",
  ai = TRUE,
  size_dots = 3,
  col_bar = TRUE,
  tf = FALSE,
  ...
)
}
\arguments{
\item{mfd}{data.table with molecular formula data as derived from
\code{ume::assign_formulas}. Column names of elements/isotopes must match names in
the \code{isotope} column of \code{ume::masses}; values are integers representing
counts per formula.}

\item{z_var}{Character. Column name for variable used for color-coding. Content of column should be numeric.}

\item{projection}{If TRUE, median z-values per (oc, hc) are used.}

\item{palname}{Character. Name of the palette. Available palettes:
\code{"black"}, \code{"redblue"}, \code{"ratios"}, \code{"rainbow"}, \code{"awi"},
\code{"viridis"}, \code{"inferno"}, \code{"terrain.colors"}, \code{"gray"}.}

\item{median_vK}{Add median VK point.}

\item{col_median}{Color of the marker for the median O/C and H/C value (Default = "white")}

\item{ai}{Add aromaticity index threshold lines.}

\item{size_dots}{Numeric. Size of the dots in the plot (default = 0.5).}

\item{col_bar}{Logical. If \code{TRUE}, adds a color legend (default is \code{TRUE}).}

\item{tf}{Logical. If \code{TRUE}, applies a transformation to the color scale (default is \code{FALSE}).}

\item{...}{
  Arguments passed on to \code{\link[=f_colorz]{f_colorz}}
  \describe{
    \item{\code{z}}{Numeric vector. Values whose colors should be computed.}
    \item{\code{col_num}}{Integer. Number of colors in the palette (default: \code{100}).}
    \item{\code{verbose}}{logical; if \code{TRUE}, show progress messages.}
  }}
}
\value{
ggplot or plotly object
}
\description{
Creates a Van Krevelen diagram (H/C vs O/C).
}
\details{
Plot Van Krevelen Diagram
}
\seealso{
Other uplots: 
\code{\link{uplot_cluster}()},
\code{\link{uplot_cvm}()},
\code{\link{uplot_dbe_minus_o_freq}()},
\code{\link{uplot_dbe_vs_c}()},
\code{\link{uplot_dbe_vs_ma}()},
\code{\link{uplot_dbe_vs_o}()},
\code{\link{uplot_freq_ma}()},
\code{\link{uplot_freq_vs_ppm}()},
\code{\link{uplot_hc_vs_m}()},
\code{\link{uplot_heteroatoms}()},
\code{\link{uplot_isotope_precision}()},
\code{\link{uplot_kmd}()},
\code{\link{uplot_lcms}()},
\code{\link{uplot_ma_vs_mz}()},
\code{\link{uplot_ms}()},
\code{\link{uplot_n_mf_per_sample}()},
\code{\link{uplot_pca}()},
\code{\link{uplot_ratios}()},
\code{\link{uplot_reproducibility}()},
\code{\link{uplot_ri_vs_sample}()}
}
\concept{uplots}
